Disable hostname from DHCP hack for Digia location in Ubuntu 10.04.
We are not relying on DHCP for providing hostnames for CI testers,
instead we customize the Virtual Machines during deployment so
that they have correct host name when booted. Now this DHCP hack
apparently breaks our system...

+++ puppet/modules/baselayout/manifests/ubuntu.pp
@@ -12,33 +12,35 @@ class baselayout::ubuntu inherits baselayout::linux {
             ;
         }
 
-        # Hostname will only be set via DHCP if /etc/hostname is absent.
-        # Gleaned from https://bugs.launchpad.net/ubuntu/+source/dhcp3/+bug/90388
-        # Therefore, we remove /etc/hostname iff resolving our fqdn does not
-        # result in the correct IP address - we should then get the correct fqdn on
-        # next boot.
-        # Note: we used to remove this file at every boot, but that seems to make the
-        # boot sequence a bit racy, as some services don't like it when the hostname
-        # changes after they've started.
-        exec { "/bin/rm -f /etc/hostname":
-            onlyif => "/bin/sh -c \"[ x$(host $::fqdn | sed -r -n -e 's/^.*has address //p') != x$::ipaddress ]\""
-        }
+        if $location != "Digia" {
+            # Hostname will only be set via DHCP if /etc/hostname is absent.
+            # Gleaned from https://bugs.launchpad.net/ubuntu/+source/dhcp3/+bug/90388
+            # Therefore, we remove /etc/hostname iff resolving our fqdn does not
+            # result in the correct IP address - we should then get the correct fqdn on
+            # next boot.
+            # Note: we used to remove this file at every boot, but that seems to make the
+            # boot sequence a bit racy, as some services don't like it when the hostname
+            # changes after they've started.
+            exec { "/bin/rm -f /etc/hostname":
+                onlyif => "/bin/sh -c \"[ x$(host $::fqdn | sed -r -n -e 's/^.*has address //p') != x$::ipaddress ]\""
+            }
 
-        # Deployment of these files forces the boot process to stall until the correct hostname
-        # has been set.  Various services don't like it if the hostname changes after they have
-        # been launched.
-        file {
-            "/etc/default/maybe_wait_for_real_hostname":
-                owner   =>  "root",
-                mode    =>  0444,
-                source  =>  "puppet:///modules/baselayout/ubuntu/maybe_wait_for_real_hostname",
-            ;
-            "/etc/default/locale":
-                owner   =>  "root",
-                mode    =>  0444,
-                source  =>  "puppet:///modules/baselayout/ubuntu/locale",
-                require =>  File["/etc/default/maybe_wait_for_real_hostname"],
-            ;
+            # Deployment of these files forces the boot process to stall until the correct hostname
+            # has been set.  Various services don't like it if the hostname changes after they have
+            # been launched.
+            file {
+                "/etc/default/maybe_wait_for_real_hostname":
+                    owner   =>  "root",
+                    mode    =>  0444,
+                    source  =>  "puppet:///modules/baselayout/ubuntu/maybe_wait_for_real_hostname",
+                ;
+                "/etc/default/locale":
+                    owner   =>  "root",
+                    mode    =>  0444,
+                    source  =>  "puppet:///modules/baselayout/ubuntu/locale",
+                    require =>  File["/etc/default/maybe_wait_for_real_hostname"],
+                ;
+            }
         }
 
         if $baselayout::testuser {
